Add feature-based similarity API and use it for images search
