+++ catalogue/webapp/components/PaletteSimilarityBox/PaletteSimilarityBox.js
@@ -1,8 +1,8 @@
 // @flow
 import { type Work } from '@weco/common/model/work';
 import { useState, useEffect } from 'react';
-import type { SimilarImage } from '../../services/labs/palette-api';
-import { getSimilarPaletteImages } from '../../services/labs/palette-api';
+import type { SimilarImage } from '../../services/labs/image-similarity';
+import { getSimilarImages } from '../../services/labs/image-similarity';
 
 type Props = {|
   work: Work,
@@ -12,7 +12,7 @@ const PaletteSimilarityBox = ({ work }: Props) => {
   const [similar, setSimilar] = useState<SimilarImage[]>([]);
 
   async function getSimilar(id) {
-    setSimilar(await getSimilarPaletteImages(id));
+    setSimilar(await getSimilarImages({ id, metric: 'palette' }));
   }
 
   useEffect(() => {
